Modified prompts to act as more of a "financial advisor"
By changing the model from 3.5 turbo to 4, and finetuning the model to have

---
FUTURE IMPROVEMENTS
---

Change "GPT is typing" to be at the bottom instead of top of the conversation. Perhaps have a bubble with 3 dots (like iMessage?) to make the user feel like it's 'thinking'?

Interpret the GPT response to look nicer. For example **Growth Stocks** should be bolded when presenting said information.

Slowly load in the response to make it feel "smoother". For example, have the information fade-in like with Gemini.

Make the size of GPT financial services container bigger vertically wise.

Save the conversation/acceptance button?

Have a button for the conversation to restart?