Can clear conversation
The user can reset the conversation through the use of a Navbar.

---
FUTURE IMPROVEMENTS
---
Interpret the GPT response to look nicer. For example **Growth Stocks** should be bolded when presenting said information.

Slowly load in the response to make it feel "smoother". For example, have the information fade-in like with Gemini.

Make the size of GPT financial services container bigger vertically wise.

Save the conversation/acceptance button?

Have an animated rainbow border for when the server is processing request.