Mobile fixes!
There is now a sticky bottom input group for mobile.

Furthermore, the chatlog goes all the way to the bottom of the screen for mobile!

---
FUTURE IMPROVEMENTS
---
Make the size of GPT financial services container bigger vertically wise.

Save the conversation/acceptance button?

Have an animated rainbow border for when the server is processing request (only on desktop).