Added local storage!
Issue: If you reload the page, and ask a follow up question, it doesn't work. From what i've noticed, this happens because the user is creating a new network connection to the server, so the previous history is lost. Perhaps, save the messages locally, and everything to the Flask sever?